refactor(furniture.models): split models into separate modules
Divide the furniture.models into individual modules to enhance code readability and maintainability. This structural change facilitates easier navigation and better scalability of the module architecture.
