+++ furniture/models/__init__.py
@@ -1,0 +1,5 @@
+from .base import Coordinate, PlacementCoordinates
+from .placements import (DoorPlacement, FurniturePlacement,
+                         PowerSocketPlacement, WindowPlacement)
+from .room_items import Furniture, RoomType
+from .room_layout import RoomLayout
